+++ src/flowDraw.js
@@ -1,22 +1,24 @@
-//import utils from './utils'
 
-const canvas = document.querySelector('canvas')
-const c = canvas.getContext('2d')
 
-canvas.width = innerWidth
-canvas.height = innerHeight/2
+import Menu from './js/menu'
+import GlobalVariables from './js/globalvariables'
+import Molecule from './js/molecules/molecule.js'
 
-//put some content in the sidebar at launch
-let sideBar = document.querySelector('.sideBar');
+GlobalVariables.canvas = document.querySelector('canvas')
+GlobalVariables.c = GlobalVariables.canvas.getContext('2d')
+
+GlobalVariables.canvas.width = innerWidth
+GlobalVariables.canvas.height = innerHeight/2
+
 let lowerHalfOfScreen = document.querySelector('.flex-parent');
 lowerHalfOfScreen.setAttribute("style","height:"+innerHeight/2.1+"px");
 
 // Event Listeners
 let flowCanvas = document.getElementById('flow-canvas');
-flowCanvas.addEventListener('contextmenu', showmenu);
+flowCanvas.addEventListener('contextmenu', Menu.showmenu); //redirect right clicks to show the menu
 
 flowCanvas.addEventListener('mousemove', event => {
-    currentMolecule.nodesOnTheScreen.forEach(molecule => {
+    GlobalVariables.currentMolecule.nodesOnTheScreen.forEach(molecule => {
         molecule.clickMove(event.clientX,event.clientY);        
     });
 })
@@ -25,9 +27,9 @@ window.addEventListener('resize', event => {
     
     console.log("resize");
     
-    var bounds = canvas.getBoundingClientRect();
-    canvas.width = bounds.width;
-    canvas.height = bounds.height; 
+    var bounds = GlobalVariables.canvas.getBoundingClientRect();
+    GlobalVariables.canvas.width = bounds.width;
+    GlobalVariables.canvas.height = bounds.height; 
 
 })
 
@@ -36,14 +38,14 @@ flowCanvas.addEventListener('mousedown', event => {
     
     var clickHandledByMolecule = false;
     
-    currentMolecule.nodesOnTheScreen.forEach(molecule => {
+    GlobalVariables.currentMolecule.nodesOnTheScreen.forEach(molecule => {
         if (molecule.clickDown(event.clientX,event.clientY) == true){
             clickHandledByMolecule = true;
         }
     });
     
     if(!clickHandledByMolecule){
-        currentMolecule.backgroundClick();
+        GlobalVariables.currentMolecule.backgroundClick();
     }
     
     //hide the menu if it is visible
@@ -58,7 +60,7 @@ flowCanvas.addEventListener('dblclick', event => {
     
     var clickHandledByMolecule = false;
     
-    currentMolecule.nodesOnTheScreen.forEach(molecule => {
+    GlobalVariables.currentMolecule.nodesOnTheScreen.forEach(molecule => {
         if (molecule.doubleClick(event.clientX,event.clientY) == true){
             clickHandledByMolecule = true;
         }
@@ -71,7 +73,7 @@ flowCanvas.addEventListener('dblclick', event => {
 
 flowCanvas.addEventListener('mouseup', event => {
     //every time the mouse button goes up
-    currentMolecule.nodesOnTheScreen.forEach(molecule => {
+    GlobalVariables.currentMolecule.nodesOnTheScreen.forEach(molecule => {
         molecule.clickUp(event.clientX,event.clientY);      
     });
 })
@@ -79,7 +81,7 @@ flowCanvas.addEventListener('mouseup', event => {
 window.addEventListener('keydown', event => {
     //every time the mouse button goes up
     
-    currentMolecule.nodesOnTheScreen.forEach(molecule => {
+    GlobalVariables.currentMolecule.nodesOnTheScreen.forEach(molecule => {
         molecule.keyPress(event.key);      
     });
 })
@@ -87,38 +89,8 @@ window.addEventListener('keydown', event => {
 
 // Implementation
 
-var availableTypes = {
-    circle:        {creator: Circle, atomType: "Circle"},
-    rectangle:     {creator: Rectangle, atomType: "Rectangle"},
-    shirinkwrap:   {creator: ShrinkWrap, atomType: "ShrinkWrap"},
-    translate:     {creator: Translate, atomType: "Translate"},
-    regularPolygon:{creator: RegularPolygon, atomType: "RegularPolygon"},
-    extrude:       {creator: Extrude, atomType: "Extrude"},
-    scale:         {creator: Scale, atomType: "Scale"},
-    intersection:  {creator: Intersection, atomType: "Intersection"},
-    difference:    {creator: Difference, atomType: "Difference"},
-    costant:       {creator: Constant, atomType: "Constant"},
-    equation:      {creator: Equation, atomType: "Equation"},
-    molecule:      {creator: Molecule, atomType: "Molecule"},
-    input:         {creator: Input, atomType: "Input"},
-    readme:        {creator: Readme, atomType: "Readme"},
-    rotate:        {creator: Rotate, atomType: "Rotate"},
-    mirror:        {creator: Mirror, atomType: "Mirror"},
-    githubmolecule:{creator: GitHubMolecule, atomType: "GitHubMolecule"},
-    union:         {creator: Union, atomType: "Union"}
-}
-
-var secretTypes = {
-    output:        {creator: Output, atomType: "Output"}
-}
-
-
-let currentMolecule;
-let topLevelMolecule;
-let menu;
-
 function init() {
-    currentMolecule = new Molecule({
+    GlobalVariables.currentMolecule = new Molecule({
         x: 0, 
         y: 0, 
         topLevel: true, 
@@ -127,23 +99,6 @@ function init() {
         uniqueID: generateUniqueID()
     });
     
-    menu = document.querySelector('.menu');
-    menu.classList.add('off');
-    menuList = document.getElementById("menuList");
-    
-    //Add the search bar to the list item
-    
-    for(var key in availableTypes) {
-        var newElement = document.createElement("LI");
-        var instance = availableTypes[key];
-        var text = document.createTextNode(instance.atomType);
-        newElement.setAttribute("class", "menu-item");
-        newElement.setAttribute("id", instance.atomType);
-        newElement.appendChild(text); 
-        menuList.appendChild(newElement); 
-        
-        document.getElementById(instance.atomType).addEventListener('click', placeNewNode);
-    }
 }
 
 function generateUniqueID(){
@@ -161,10 +116,9 @@ function distBetweenPoints(x1, x2, y1, y2){
 // Animation Loop
 function animate() {
     requestAnimationFrame(animate);
-    c.clearRect(0, 0, canvas.width, canvas.height);
+    GlobalVariables.c.clearRect(0, 0, GlobalVariables.canvas.width, GlobalVariables.canvas.height);
     
-    //c.fillText('T', mouse.x, mouse.y)
-    currentMolecule.nodesOnTheScreen.forEach(molecule => {
+    GlobalVariables.currentMolecule.nodesOnTheScreen.forEach(molecule => {
         molecule.update();
     });
 }
