+++ src/js/molecules/regularPolygon.js
@@ -1,4 +1,6 @@
-class RegularPolygon extends Atom {
+import Atom from '../prototypes/atom'
+
+export default class RegularPolygon extends Atom {
 
     constructor(values){
         super(values)
