+++ src/js/molecules/shrinkwrap.js
@@ -1,4 +1,6 @@
-class ShrinkWrap extends Atom{
+import Atom from '../prototypes/atom'
+
+export default class ShrinkWrap extends Atom{
     
     constructor(values){
         super(values);
