+++ webpack.config.js
@@ -35,5 +35,5 @@ module.exports = {
         })
     ],
     watch: true,
-    devtool: 'source-map'
+    devtool: 'inline-source-map'
 }
